requirement.txt
beautifulsoup4==4.6.0
certifi==2018.4.16
chardet==3.0.4
idna==2.7
Pillow==5.2.0
pycrypto==2.6.1
PyExecJS==1.5.1
requests==2.19.1
six==1.11.0
tqdm==4.23.4
urllib3==1.23
